+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove `RSpec/FilePath` cop. ([@ydah])
 - Add new `RSpec/ExpectInLet` cop. ([@yasu551])
 - Remove `RSpec/Capybara/FeatureMethods` cop. If you are using this cop, change it to use `RSpec/Dialect`. ([@ydah])
+- Add new `RSpec/ExpectationTargetMethod` cop. ([@krororo])
 
 ## 2.29.2 (2024-05-02)
 
@@ -922,6 +923,7 @@ Compatibility release so users can upgrade RuboCop to 0.51.0. No new features.
 [@k-s-a]: https://github.com/K-S-A
 [@kellysutton]: https://github.com/kellysutton
 [@koic]: https://github.com/koic
+[@krororo]: https://github.com/krororo
 [@kuahyeow]: https://github.com/kuahyeow
 [@lazycoder9]: https://github.com/lazycoder9
 [@leoarnold]: https://github.com/leoarnold